pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataflow-agent"
version = "0.1.0"
description = "Intelligent data governance and automated cleaning engine based on multi-agent collaboration"
requires-python = ">=3.11"
dependencies = [
    # Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    # Data models
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    # Async task queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    # Data processing
    "polars>=0.20.31",
    "pandas>=2.2.2",
    "pyarrow>=15.0.0",
    # Entity resolution
    "splink>=3.9.7",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    # Config / utils
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    # Sandbox
    "docker>=7.0.0",
    # Observability
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "prometheus-client>=0.20.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
dataflow-run   = "scripts.run_job:cli"
dataflow-eval  = "scripts.evaluate:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["dataflow*", "scripts*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
